Decoded argument values are now being correctly added to Instruction (instance) objects - all Level4 compiler warnings fixed

+++ src/Instruction.cpp
@@ -6,7 +6,7 @@ Instruction::Instruction(){
 	binString = "";
 	bin = 0x0;
 }
-Instruction::Instruction(InstructionData* id, string asmString, string binString, instr bin, vector<int64_t> arguments)
+Instruction::Instruction(InstructionData* id, string asmString, string binString, instr bin, vector<int> arguments)
 	:id(id), asmString(asmString), binString(binString), bin(bin), arguments(arguments)
 {
 
@@ -21,7 +21,7 @@ InstructionData* Instruction::getId(){return id;}
 string Instruction::getAsmString(){return asmString;}
 string Instruction::getBinString(){return binString;}
 instr Instruction::getBin(){return bin;}
-vector<int64_t> Instruction::getArguments(){return arguments;}
+vector<int> Instruction::getArguments(){return arguments;}
